+++ integrations/amazon_bedrock/src/haystack_integrations/common/amazon_bedrock/__init__.py
