+++ integrations/amazon_bedrock/src/haystack_integrations/common/amazon_bedrock/errors.py
