+++ integrations/amazon_bedrock/src/haystack_integrations/components/generators/amazon_bedrock/generator.py
@@ -7,11 +7,11 @@ from botocore.exceptions import ClientError
 from haystack import component, default_from_dict, default_to_dict
 from haystack.utils.auth import Secret, deserialize_secrets_inplace
 
-from haystack_integrations.commons.amazon_bedrock.errors import (
+from haystack_integrations.common.amazon_bedrock.errors import (
     AmazonBedrockConfigurationError,
     AmazonBedrockInferenceError,
 )
-from haystack_integrations.commons.amazon_bedrock.utils import get_aws_session
+from haystack_integrations.common.amazon_bedrock.utils import get_aws_session
 
 from .adapters import (
     AI21LabsJurassic2Adapter,
